+++ rascil/processing_components/imaging/__init__.py
@@ -31,3 +31,4 @@ from .dft import *
 from .ng import *
 from .imaging import *
 from .imaging_helpers import *
+from .ao import *