+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 master
 ------
+* Add ao.py which directly copied a few AOTools functions ([MR306](https://gitlab.com/ska-telescope/external/rascil/-/merge_requests/306))
 * Remove Numba related code ([MR303](https://gitlab.com/ska-telescope/external/rascil/-/merge_requests/303))
 * Install PyBDSF from PyPi test_skymodel_extract_skycomponentsusing latest python3.9 wheel ([MR300](https://gitlab.com/ska-telescope/external/rascil/-/merge_requests/300))
 * Bug fix in initial SkyModel for ICAL: convert polarisation of SkyComponents read from file to image polarisation ([MR296](https://gitlab.com/ska-telescope/external/rascil/-/merge_requests/296))